Adding support for Auto-Updating. http://code.google.com/p/videobrowser/issues/detail?id=14
Updater.cs was added to the Util package that handles the updating.

Application.cs was updated with 2 modifications:
1) Addition of spawning the update thread (currently a 1 time check) on load.
2) Support for thread-safe display of dialog messages.  (Previously a dialog message from another thread could get lost on a page navigation as they are deferred callbacks).

Config.cs was updated to support 2 new features:
EnableUpdates - Defaults to true, enables checking for updates.
EnableBetas - Defaults to false, enables update checking to consider beta versions.

DevInstall.cmd was updated to fix the gacutil call in the uninstall that was incorrect.

the .csproj file was updated for the new Updater.cs file.

The Updater.cs file has a string in it that must still be replaced with the location of the version info.xml file.  The XML should be of the following format:

<FileInfo>
 <betaVersion>200</betaVersion>
 <betaURL>http://thelucks.org/vb/setup.msi</betaURL>
 <ver>200</ver>
 <url>http://videobrowser.googlecode.com/files/MCE_Video_Browser_1_1_13_0.msi</url>
</FileInfo>


There are additional actions to take now when creating releases:
1) Update the Version member in the Application class. (int value)
2) Update the information in the version information XML file.