- External player support, add section to config file for <ExternalPlayers>   <ExternalPlayer>    <MediaType>Avi</MediaType>    <Command>wmplayer.exe</Command>    <Args>"{0}"</Args>   </ExternalPlayer> </ExternalPlayers>
Where ExternalPlayer element is repeated for each required mediaType.
Valid MediaType values (case sensitive)are Avi, Mkv, DVD, BlueRay, HDDVD
The {0} in the args parameter will be replace by the name of the item that is being played - i.e. the folder name but excluding the video_ts bit etc.